Implement: Update processor and provider function signatures [#255]
Updates processor and provider functions to use specialized identifier types:
- world/processor.go: worldId byte → world.Id
- world/requests.go: worldId byte → world.Id
- data/map/processor.go: remove cast, use _map.Id directly
- data/map/requests.go: mapId uint32 → _map.Id
- guild/model.go: worldId byte → world.Id with WorldId() accessor
- guild/rest.go: convert byte to world.Id in Extract()
- kafka/consumer/session/consumer.go: remove redundant cast

+++ atlas.com/channel/data/map/processor.go
@@ -25,5 +25,5 @@ func NewProcessor(l logrus.FieldLogger, ctx context.Context) Processor {
 }
 
 func (p *ProcessorImpl) GetById(mapId _map.Id) (Model, error) {
-	return requests.Provider[RestModel, Model](p.l, p.ctx)(requestMap(uint32(mapId)), Extract)()
+	return requests.Provider[RestModel, Model](p.l, p.ctx)(requestMap(mapId), Extract)()
 }
